# app/Http/Resources/Auth/RegisterResource.php
<?php

namespace App\Http\Resources\Auth;

use Illuminate\Http\Resources\Json\JsonResource;

class RegisterResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @param \Illuminate\Http\Request $request
     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
     */
    public function toArray($request)
    {
        return [
            'user' => [
                'id' => $this->id,
                'name' => $this->name,
                'email' => $this->email,
                'email_verified_at' => $this->email_verified_at ?? '',
                'created_at' => $this->created_at,
                'updated_at' => $this->updated_at
            ],
            'access_token' => $this->createToken('authToken')->accessToken
        ];
    }
}